Crate spc700

Implements the Audio Processing Unit (APU)

The APU consists of 64 KB shared RAM, the SPC700 and the DSP. It is almost fully independent from the rest of the SNES.

The SPC700 is an independent audio coprocessor. The main CPU can transmit a program and audio data into the shared RAM and then execute it. The program can manipulate DSP registers and specify samples to play.

Structs

Spc700
The SPC700 is an 8-bit processor with a 16-bit address space.