Struct spc700::Spc700

pub struct Spc700 {
    pub trace: bool,
    // some fields omitted
}

The SPC700 is an 8-bit processor with a 16-bit address space.

It has 64 KB of RAM shared with the DSP. The last 64 Bytes in its address space are mapped to the "IPL ROM", which contains a small piece of startup code that allows the main CPU to transfer a program to the APU.

Fields
trace: bool
Methods

impl Spc700

pub fn store_port(&mut self, port: u8, value: u8)

Store a byte in an IO port (0-3)

SNES IO ports $2140-$2143 are mapped to internal registers $f4-$f7

pub fn read_port(&mut self, port: u8) -> u8

Load a byte from an IO port

pub fn dispatch(&mut self) -> u8

Dispatch an opcode